Introduce caching for InvokeOperations
